% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arrow_head_default}
\alias{arrow_head_default}
\title{Does nothing but provide defaults for documentation}
\usage{
arrow_head_default(
  rotate = 0,
  rescale = c(1, 1),
  nudge = c(0, 0),
  transformations = c("unitizer", "rotater", "rescaler", "nudger"),
  plot = FALSE
)
}
\arguments{
\item{rotate}{rotation angle in radians}

\item{rescale}{a single value or 2-length vector for scaling in x and y}

\item{nudge}{a single value or 2-length vector for nudging in x and y}

\item{transformations}{a vector of transformation functions}

\item{plot}{plot arrowhead if TRUE}

\item{n}{number of points in polygon}
}
\value{
a matrix
}
\description{
Does nothing but provide defaults for documentation
}
\keyword{internal}
