% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arrow_head_harpoon}
\alias{arrow_head_harpoon}
\title{Make a harpoon arrowhead}
\usage{
arrow_head_harpoon(
  point_angle = 30,
  barb_angle = 20,
  degrees = TRUE,
  rotate = 0,
  rescale = c(1, 1),
  nudge = c(0, 0),
  transformations = c("unitizer", "rotater", "rescaler", "nudger"),
  plot = FALSE
)
}
\arguments{
\item{point_angle}{angle of harpoon point}

\item{barb_angle}{angle of harpoon barb}

\item{degrees}{if TRUE, angles are in degrees instead of radians}

\item{rotate}{rotation angle in radians}

\item{rescale}{a single value or 2-length vector for scaling in x and y}

\item{nudge}{a single value or 2-length vector for nudging in x and y}

\item{transformations}{a vector of transformation functions}

\item{plot}{plot arrowhead if TRUE}
}
\value{
a matrix
}
\description{
Make a harpoon arrowhead
}
\examples{
arrow_head_harpoon(plot = TRUE)
}
