% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arrow_head_icon}
\alias{arrow_head_icon}
\title{Make arrowhead from preset icon}
\usage{
arrow_head_icon(
  x = "stardestoyer",
  rotate = 0,
  rescale = c(1, 1),
  nudge = c(0, 0),
  transformations = c("rotater", "rescaler", "nudger"),
  plot = FALSE
)
}
\arguments{
\item{x}{name of icon: eiffel, viper, viper2, nighthawk, pantherxf70}

\item{rotate}{rotation angle in radians}

\item{rescale}{a single value or 2-length vector for scaling in x and y}

\item{nudge}{a single value or 2-length vector for nudging in x and y}

\item{transformations}{a vector of transformation functions}

\item{plot}{plot arrowhead if TRUE}
}
\value{
a matrix
}
\description{
Make arrowhead from preset icon
}
\examples{
starwars_stardestoyer <- arrow_head_icon(x = "stardestoyer", plot = TRUE)
starwars_executor <- arrow_head_icon(x = "executor", plot = TRUE)
eiffel <- arrow_head_icon(x = "eiffel", plot = TRUE)
battlestar_galactica_viper <- arrow_head_icon(x = "viper", plot = TRUE)
battlestar_galactica_viper2 <- arrow_head_icon(x = "viper2", plot = TRUE)
nighthawk <- arrow_head_icon(x = "nighthawk", plot = TRUE)
panther_xf70 <- arrow_head_icon(x = "pantherxf70", plot = TRUE)
}
