% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arrow_head_trefoil}
\alias{arrow_head_trefoil}
\title{Make trefoil arrowhead}
\usage{
arrow_head_trefoil(
  rotate = 0,
  rescale = c(1, 1),
  nudge = c(0, 0),
  transformations = c("unitizer", "rotater", "rescaler", "nudger"),
  n = 361,
  plot = FALSE
)
}
\arguments{
\item{rotate}{rotation angle in radians}

\item{rescale}{a single value or 2-length vector for scaling in x and y}

\item{nudge}{a single value or 2-length vector for nudging in x and y}

\item{transformations}{a vector of transformation functions}

\item{n}{number of points in polygon}

\item{plot}{plot arrowhead if TRUE}
}
\value{
a matrix
}
\description{
Make trefoil arrowhead
}
\examples{
trefoil <- arrow_head_trefoil(plot = TRUE)
}
