% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{arrowhead_ellipse}
\alias{arrowhead_ellipse}
\title{Make arrrowhead with ellipse}
\usage{
arrowhead_ellipse(
  a = 1,
  b = 1,
  nudge = c(0, 0),
  rescale = 1,
  rotate = 0,
  n = 361,
  plot = FALSE
)
}
\arguments{
\item{a}{width of ellipse}

\item{b}{height of ellipse}

\item{nudge}{length 2 vector for nudging x and y}

\item{rescale}{length 2 vector for rescaling x and y}

\item{rotate}{radians for rotating}

\item{n}{number of points in ellipse}

\item{plot}{plot arrowhead}
}
\value{
a matrix
}
\description{
Make arrrowhead with ellipse
}
\examples{
arrowhead_ellipse(plot = T, b = .5)
arrowhead_ellipse(
  plot = T,
  b = .5,
  rescale = .45,
  nudge = c(.55, 0)
)
}
