% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bezier2matrix}
\alias{bezier2matrix}
\title{make arrowhead from list of bezier control points}
\usage{
bezier2matrix(x, t = seq(0, 1, 0.01), plot = FALSE)
}
\arguments{
\item{x}{list of control points (as vectors or matrices)}

\item{t}{sequence of points (usually 0 to 1)}

\item{plot}{plot arrowhead}
}
\value{
a matrix
}
\description{
make arrowhead from list of bezier control points
}
\examples{
bezier2matrix(list(c( 1,  0,
                      1, .5,
                     .2,  1),
                   c(.2,  1,
                    -.5, .5,
                      0,  0)),
              plot = TRUE)
}
