% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nudger}
\alias{nudger}
\title{Nudge columns of a matrix by fixed amounts}
\usage{
nudger(x, nudge)
}
\arguments{
\item{x}{a matrix}

\item{nudge}{a single value or a vector with length equal to the number of columns in x}
}
\value{
matrix
}
\description{
Nudge columns of a matrix by fixed amounts
}
\examples{
nudger(matrix(0, nrow = 2, ncol = 2), nudge = c(0,1))
}
