% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_arrowhead}
\alias{plot_arrowhead}
\title{Plot arrowhead}
\usage{
plot_arrowhead(x, displayline = TRUE, displaypoints = TRUE)
}
\arguments{
\item{x}{2-column matrix}

\item{displayline}{plot the display line}

\item{displaypoints}{plot the 0,0 point and the 1,0 point}
}
\value{
plot
}
\description{
Plot arrowhead
}
