% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reflecter}
\alias{reflecter}
\title{make a reflection of a matrix on the y axis}
\usage{
reflecter(x, add_reflection = TRUE)
}
\arguments{
\item{x}{matrix}

\item{add_reflection}{add to x in reverse order}
}
\value{
a matrix with y reversed sign and rows in reverse order
}
\description{
Good for making symmetrical arrowheads
}
\examples{
reflecter(diag(c(1,2)))
}
