% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rev_matrix_rows}
\alias{rev_matrix_rows}
\alias{rev_matrix_cols}
\title{reverses the order of rows or columns in a matrix}
\usage{
rev_matrix_rows(x)

rev_matrix_cols(x)
}
\arguments{
\item{x}{matrix}
}
\value{
a matrix
}
\description{
reverses the order of rows or columns in a matrix
}
\examples{
rev_matrix_rows(diag(c(1,2)))
}
