% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rotater}
\alias{rotater}
\title{Rotate a 2-column matrix}
\usage{
rotater(x, theta, center = c(0, 0), degrees = FALSE)
}
\arguments{
\item{x}{a 2-column matrix}

\item{theta}{angle}

\item{center}{point of rotation}

\item{degrees}{if TRUE, theta is in degrees instead of radians}
}
\value{
a rotated 2-column matrix
}
\description{
Rotate a 2-column matrix
}
\examples{
x <- matrix(seq(10), ncol = 2)
rotater(x, pi)
}
