% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_arrowhead}
\alias{plot_arrowhead}
\title{Plot arrowhead}
\usage{
plot_arrowhead(x, displayline = TRUE, displaypoints = TRUE)
}
\arguments{
\item{x}{2-column matrix}
}
\value{
plot
}
\description{
Plot arrowhead
}
