% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{unitizer}
\alias{unitizer}
\title{Fit matrix to unit circle}
\usage{
unitizer(x, center = rep(0, ncol(x)))
}
\arguments{
\item{x}{matrix}

\item{center}{center of matrix}
}
\value{
matrix
}
\description{
Fit matrix to unit circle
}
\examples{
A = matrix(c(1, 2,
            -8,6,
             9,5),
            ncol = 2,
            byrow = TRUE)
unitizer(A)
cA <- unitizer(A, center = colMeans(A))
plot(cA, xlim = c(-1, 1), ylim = c(-1, 1))
t <- seq(0,2*pi, length.out = 361)
lines(cos(t), sin(t))
}
