% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{v2matrix}
\alias{v2matrix}
\title{Convert a vector to a matrix}
\usage{
v2matrix(x, ncol = 2, byrow = TRUE)
}
\arguments{
\item{x}{vector}

\item{ncol}{number of columns}

\item{byrow}{logical. convert by row}
}
\description{
Convert a vector to a matrix
}
